Workaround to the sharp module bug.
When trying to set up fCC locally without Docker, many people have been experiencing errors relating to the `sharp` module while issuing the `npm run develop` command. A quick fix to this is being passed down from mouth to mouth in the forum and in gitter, but it hasn't made its way into the official docs. It's time for it to do so, since this frustrates many attempts for would-be contributors to start contributing to the fCC project.