chore: add formatter (fcc md to standard gfm)
The current parser behaves differently if there is a blank line after
the section tag for instructions and descriptions.  formatCurriculum
rewrites the challenges into GFM and inserts a space after the section
tags to ensure the md is parsed correctly.

In addition it formats the markdown and replaces code tags with
backticks to improve readability.

To improve bare url handling and escaping wrapBareUrls is localised so
that it only changes the bare urls (including urls in html), not the
surrounding markdown.

Entities are convertd into characters when escaping, so the challenge
parser will interpret them correctly.

Html for dfn, sup, sub and similar are kept. This is in contrast to the
default conversion to markdown which discards those tags.