Fix link to latest unanswered posts in the forum
The link to the forums had a `max_posts=1` argument, but this does not work on the home page that it was applied to. 

Instead, we should use `/latest?max_posts=1` which this commit does. 

If, instead, we wish to send users to the home page of the forums, then removing the `max_posts=1` argument would also make this link correct.