revised to effectively eliminate all processes
I included a while loop to effectively loop through all of the processes. In that way, everything gets eliminated. I tested this script locally on my machine with local variables of my creation. However, I was unable to reproduce the kill signals on my machine. Thank you so much for taking the time to review my pull request!!