Clarify instructions to remove confusion
I found the instructions as written very confusing. The first sentence tells us to "Use the `playerNumber` variable to look up player 16 in `testObj` using bracket notation." This seems to suggest that setting the `playerNumber` variable in the first place somehow involves `testObj`. But it does not. The `playerNumber` variable is simply set to the number 16. It is only when `player` gets a value that we do anything with `testObj`. The point of this exercise is to show how the reader can assign a key (16) to a variable and then look up the key's associated value by using the variable rather than the key itself. I think that this comes across much more clearly, without confusion, in my proposed revised instruction.

Note: I can imagine someone saying "Well, we want the reader to figure it out, so it's okay if the instructions are confusing." I disagree. We shouldn't be challenging the reader to understand confusing instructions. In general, the FreeCodeCamp instructions are very clear. This problem is just an exception, and it is easy enough to make it clear without losing the fundamental point of the exercise.