Clarify when bracket notation is necessary
Currently, the exercises about objects do not explain that you must use bracket notation when accessing a property by a variable name. This creates confusion in a later exercise, https://www.freecodecamp.org/learn/javascript-algorithms-and-data-structures/basic-javascript/testing-objects-for-properties, because it fails if you use dot notation, and there's no explanation about this. This has confused not just me but also other users (see https://forum.freecodecamp.org/t/basic-javascript-testing-objects-for-properties/347697).

One way to remove the confusion is to include an explanation here that bracket notation is required not just when there's a space in a property name, but also when you're trying to access a property with a variable name.