Clarify how object names are associated with destructured fields
I have added some language to clarify how an object's name is associated with its fields when the object is destructured in place in a function definition.

As written, I found the examples confusing. The first example, where the object is destructured *not* in place, makes perfect sense, and it is clear that the specified fields come from the object `profileData` because the object is named in the example.

But in the second example, involving destructuring in place, the object `profileData` is not named anywhere. Instead, the example just includes field names. As I read this, I was scratching my head, asking, "How does the program know that these field names are associated with the `profileData` object? I don't see the object's name anywhere."

Of course, the answer is that the object's name gets passed as an argument when the function is called. I'm sure this is obvious to experienced programmers. But it wasn't obvious to me, and I'm an educated beginner. In fact, to figure it out, I played around with the code in repl.it for a while before I figured it out. I don't think that this is something the reader needs to figure out on his own at this point. It's not part of the pedagogical goals here. So I propose a minor revision that explains how an object's name is associated with the fields when the object is destructured in place in a function definition.