Clarify instructions about .test() method
The current instructions led me toward a solution that fails the test but that seems like it should pass. It fails because the test looks at the value of a variable called "return," not directly at the return value of myRegex.test(myString). The instructions, however, just tell you to use the .test() method, and the directions say "your result should return true." But if the last line of your code is simply <code>myRegex.test(myString)</code>, that *does* return true. And that code is equivalent to what is shown in the instructions as returning true. Really, the test is not asking for "result" to "return true"; result is not a function or a method, it is a variable. The test is really asking for "result" to *equal* true. The simplest way to solve this is to clarify that the return value of the method should be assigned to the variable <code>result</code>.