Add warning on challenge description
When I took was going through the camp, I got stuck on this challenge ("Use CSS Selectors to Style Elements"). After some checking and asking on the forums, it seems that it was not due to my code, but due to the dark mode extension I was using. I added a warning to make sure people won't get stuck on this part