Fix low-contrast CSS strings in code blocks (#39967)
First noticed on the challenge "Divide the Grid Into an Area Template", text is hard to read in night mode due to low contrast in the CSS code block. Problem was discussed in issue #39967, and this change was proposed to boost contrast above the minimum acceptable ratio of 4.5:1.

This change only targets CSS code blocks in night mode, and only strings. I could not find code blocks in other languages with this issue, but they may exist.
This also does not target `.token.operator`, `.token.entity`, or `.token.url`. There is no issue here with contrast, but these classes do not include a background. A proposed future change could be to remove the background from `.token.string` as well, but this change is not made here, as current style guides recommend the background.