tests: Modify translation process
The tests on the footer errored out if the client
was last built in a language other than English.
This was due to the values showing up as hard-
coded in the jest snapshot. I've added the trending
json files to the i18n namespace list, allowing us
to parse them through the translation function. This
means they render in the snapshot as the variable
string, and the tests pass correctly. I additionally
fixed a typo in the class name.

Signed-off-by: nhcarrigan <nhcarrigan@gmail.com>