fix(tools): validate keys in all objects
Validates the keys in all of the translation JSON objects. Note that
this will return some false positives for dynamic keys (like the
superblock intros). This is okay as the script is not meant to be
used as blocking CI, but only to help us clean up keys as we update
our client/API.