Emphasize that solution links need to be of form https:// especially for Glitch users
I found out about this the frustrating way. The URL I pasted from glitch was of the form http:// and that's why my solution was not able to pass the tests, even though it worked. 
Once I changed the URL to https:// it passed the tests. 
I think other students should be very aware of this to not cause unnecessary confusion about their solutions.