give-links-meaning-by-using-descriptive-text testing bug
The assertion for "The a element should have a closing tag" test needs the formatting to be exactly as it is written, so no extra spaces when assigning (href = ""). I don't know if the assertion was written to also make sure that the href was empty with the (''||"") part, so changing it to match the same test from link-to-external-pages-with-anchor-elements could mess something up. The contributing guidelines have ways to download and test but i figured the maintainers could read this and know what to do better.