Possible bug at the hints
At the first picture hint I was given was "The text Loving should no longer be located directly to the right of your checkbox. It should be wrapped in a label element." this. I think the hint I was supposed to be given was "The new label element should be located directly to the right of your checkbox. Make sure there is a space between the two elements." this. Because as you can see in the second picture putting space between the elements fixes the problem. I don't know if this is how I am supposed to report bugs, you are free to change anything inappropriate.