test: fix donate page for existing donors
This time we very much care about the user, so we have to have them log
in.

I removed the 'after' hook, because it's something to be avoided if
possible. It should be possible to run each Cypress test independently
and not rely on any previous test to clean up.