Contributing Chat Server - Direct Link
A link that took you straight to the chat server.  There is only one location on the README and it is not located in the contribution section.  The other location is deeper into the documentation.  Proposal is to have a link that is in the READMEs contribution section so new contributors can ask questions and see the community straight away.