fix: add tests as knip entry files
It is slightly questionable to unit test components you don't use
directly, but the justification is that we don't always need to pass in
the props that redux provides. So, this reduces the test boilerplate.

With that in mind, we don't want knip to throw false positives for
exports that only see use in tests.