test: always start server before testing
The goal is to remove boilerplate from tests by adding a common task
(starting up the server) to part of Jest's setup.

It's quite fast to start up the server ( roughly 400ms), so as long as
we don't create hundreds of suites, it should not be a big deal perf
wise.

Unfortunately it's not possible to both add the fastify instance as a
global and set it up once. At the time of writing, Jest does not support
this.