Update the navbar to properly fit the width change of the page
The navigation bar now properly adjusts based on the max-width and min-width. Prior to the change, there was an awkward state in the middle of the min-width and max-width. 

Changed the nav-logo to be in the proper position when the max-width:980px is triggered. Now, the navbar is responsive to all size.