fix(curriculum): update test question to match provided answers for TOP Question K
The test question was not related to the test answers. I replaced the question to match provided answers.

Old test question: Which of the following statements accurately describes the uses of `div` and `span` elements in HTML?

New test question from the assignment description: What happens when you add a bottom padding of 15 pixels to a `span` element with display set to `inline-block`?