fix: initialize tests on challenge mount
Even though these lessons don't use tests, it's important to reset the
tests in the redux store. If this isn't done, it's possible for another
challenge's tests to remain in the store and the new challenge will not
be able to pass them.

We considered clearing tests between challenges, but decided it was
better to make challenge mounting be responsible for challenge
store updates. It seems simpler than spreading the concern over mount
and unmount.

Co-authored-by: Tom <20648924+moT01@users.noreply.github.com>