fix: use DOMParser to parse html, not .innerHTML
We're only interested in parsing the HTML so that we can transform the
text - adding it to the DOM was a hack to simplify transformation.
DOMParser is a more direct way to achieve the same goal with fewer side
effects.

In particular, this no-longer tries to fetch any stylesheets during
parsing, so <link rel="stylesheet" href="styles.css" /> is ignored.