TYPO - Typography Mistake
Fixes #55961

Issue:
A sentence in the description is missing the word "object". It currently says:

A NodeList is an array-like, which means you can iterate through it and it shares some common methods with an array.

It should say:

A NodeList is an array-like object, which means you can iterate through it and it shares some common methods with an array.