refactor: split challenge count and donation logic
While the donation logic uses the session storage, it is not storing
anything related to donations.

It does three things: increments the completed challenge
count, saves the current count for later comparison and returns data
describing what it has stored. All of this is useful for the donation
logic, but it's quite generic.