feat(client): add worker messaging module
The normal pattern for worker communication is via onMessage listeners.
This makes for awkward code when all you really want is to be able to
write code like this:
const res = await fetch(...)

This module creates a similar api, by abstracting over the message
passing, so that the client can simply await the Promise'd response.