fix(client): render hint in jaw, not log output
Most of the time the hint and the second log are the same thing, but
not if the learner's code calls console.log twice. Then it's the value
of the second call.

This means we cannot use the output as a proxy for the hints.
Fortunately, the hints are directly accessible.

As a side benefit, this fixes a bug where the number of the test was
appearing in the hint (e.g. if it's the 5th test 5 would appear).