refactor: simplify is-audited logic
Whether or not a superblock is audited is independent of whether or not
it appears in the client. We don't need to care that the superblock is
not going to be rendered when determining if it is audited. In that case
the audited status will not be checked by the client.