fix: resize xterm with container
This only works 100% on desktop since it relies on the dimensions prop
for the height, without which it defaults to 408px. The obvious fix,
setting the height of the container to 100%, doesn't work because
xterm throws an error while trying to render the PreviewPortal.

It's not obvious to me why that's happening.