refactor: better test errors
I don't know why, but if the timeout rejects, this creates an unhandled
rejection (despite being inside a try-catch block). This happens even
if the rejection loses the race.

While we could delete the race, it gives us some information: if it
fails, we know that particular test is far too slow.