feat(GHA): run on workflow completion
This triggers a run on workflow completion for the Node.js tests which
themselves run on push. But we may just go back to using a on.push event
if this becomes a gotcha.

The idea is to only start a deployment on the prod-* branches if the
tests pass.