refactor: re-organise exam-env tools
Putting all the tooling into ./tools both makes some organisational
sense (it's not 'src', exactly) and makes it easier to ignore when it
comes to building.

While it was possible to exclude it with .dockerignore, this wasn't
desirable because it meant that the build script behaved differently
when run as part of the image build from when it is run indendently.

What ended up happening was that the __mock__ files would be included
transitively (via seed/index.ts) and fail to compile due to other files
missing.

Now we skip the mocks and the tooling whenever we build (inside and
outside of images).