feat: use version for cache busting
The reason for adding yet another way of cache busting is that the test
runner usesthe names of the evaluator files internally. That means we
cannot simply prefix them with content hashes.

In principle, when building the test runner, we could add hashes to the
evaluator bundles NOT hash the main bundle and ask the client to add a
hash to just the main bundle... but that feels much worse.