refactor: make webpack directly responsible for test runner again
This is a bit awkward, but the curriculum tests are responsible for
checking that the client would be able to run the challenge tests and so
we want the test runner to be included as part of the client. This makes
it quite natural to have the client's webpack step copy the test runner
files. However, it doesn't seem possible to execute the webpack step in
isolation and starting the client just for that is both difficult and
time consuming.

In order to make sure there's just one version of the helpers and to
allow the curriculum tests to build using that version, I went back to
using our webpack step to copy the test runner into client/static.

Unfortunately this means we have to do the declaration of
window.FCCTestRunner manually (because we're no longer directly
importing the package into the client). There may be a better way, but
I'm not seeing it.

While I was at it, I realised that the path makes more sense as
/js/test-runner/${version} than /js/${version}/test-runner.

Finally, I finished renaming FCCSandbox to FCCTestRunner.