test: relax challenge id uniqueness constraints
If the same challenge id appears twice in the same block, it's likely
something went wrong. However, we regularly reuse challenge ids in order
to give learners credit for work they've already completed.

It was possible to handle this by regularly updating the list of
duplicates, but doing so was a chore.