Create Eye to Eye
We are into the Entrepreneurship and Idea Funding Business. The state of Indian Startup Echo-systems reports 2021, India has 5694 active investors and 38k startups. 
Looking at the current size of the Market this territory shows exponential growth in coming years. The opportunity that we have identified in the funding of Businesses and startups lack Capital, Investors, Guidance, and Ideas Getting picked up.
Our Business will bridge these young entrepreneurs and investors.
We want to Capture this opportunity of making it more convenient for every entrepreneur who wants to start a business.
Our objective is to get the funding starting from 10k up to 100 crores & give hope to every new or existing entrepreneur to join Us.