fix: filter by challengeId when building blocks
We need the next challenge to get the solution, but we don't want to
test it. We just want the one specified by FCC_CHALLENGE_ID.

The issue was that filtering by challengeId was not happening. All the
block's challenges would arrive to be tested, then all but the first one
would be discarded.