Digital Pet Game Tests
Initial pass nearly complete, lingering error regarding the change of
value after excessive idle time. I've tried a few approaches and can't
get it to organically exceed what appears two be two ticks.

All other tests seems to be working as expected so far, but testing the
state of how the application responds to button clicks needs to be made
more thorough. This isn't difficult, but is cleaned with the ability to
simulate idle time.