fix: prevent eslint hoisting rather than using override
This should result in less maintenance, since the other approach would require use to keep updating the overrides.
As I understand it, this works because each eslint package should be in the workspace's node_modules so node does not
need to look in the root node_modules (which may contain incompatible versions) to find the right package.