Correct failing tests
All tests began failing as none of the helper functions in the
`before-all` block were accessible. I believe this was the result of the
challenge type being changed, DOM challenges are supposed to function
this way, but the docs on `how-to-work-on-coding-challenges` state that
other challenges need to have variables that are reused assigned to
`globalThis`.

Adding them all to `globalThis` resolved the bulk of the issues. There
were some lingering case and id issues in the sample solution, but I
updated that with my own sample code I was testing against in the
client.